+++ python/source/state_representation/bind_state.cpp
@@ -43,6 +43,7 @@ void state(py::module_& m) {
   c.def("set_empty", &State::set_empty, "Setter of the empty attribute", "empty"_a=true);
   c.def("set_filled", &State::set_filled, "Setter of the empty attribute to false and also reset the timestamp");
   c.def("get_timestamp", &State::get_timestamp, "Getter of the timestamp attribute");
+  c.def("set_timestamp", &State::set_timestamp, "Setter of the timestamp attribute");
   c.def("reset_timestamp", &State::reset_timestamp, "Reset the timestamp attribute to now");
   c.def("get_name", &State::get_name, "Getter of the name");
   c.def("set_name", &State::set_name, "Setter of the name");